requirements.txt
python-dotenv==0.19.2
mysql-connector-python==8.0.28
Jinja2==3.1.1
flamapy==1.0.1
flamapy-fm==1.0.1
flamapy-sat==1.0.1
